+++ projects/coreui-angular/src/lib/breadcrumb/breadcrumb.component.spec.ts
@@ -1,0 +1,33 @@
+import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import {CommonModule} from '@angular/common';
+import { RouterTestingModule } from '@angular/router/testing';
+import { Router } from '@angular/router';
+
+import { BreadcrumbComponent } from './breadcrumb.component';
+import { BreadcrumbService } from './breadcrumb.service';
+
+describe('BreadcrumbComponent', () => {
+  let component: BreadcrumbComponent;
+  let fixture: ComponentFixture<BreadcrumbComponent>;
+  let router: Router;
+
+  beforeEach(async(() => {
+    TestBed.configureTestingModule({
+      imports: [ CommonModule, RouterTestingModule.withRoutes([]) ],
+      declarations: [ BreadcrumbComponent ],
+      providers: [ BreadcrumbService ]
+    })
+    .compileComponents();
+  }));
+
+  beforeEach(() => {
+    fixture = TestBed.createComponent(BreadcrumbComponent);
+    router = TestBed.get(Router);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});