+++ projects/coreui-angular/src/lib/header/header.module.ts
@@ -1,0 +1,22 @@
+import { CommonModule } from '@angular/common';
+import { NgModule } from '@angular/core';
+import { RouterModule } from '@angular/router';
+
+import { LayoutModule } from '../shared/layout/layout.module';
+import { HeaderComponent } from './header.component';
+
+@NgModule({
+  imports: [
+    CommonModule,
+    RouterModule,
+    LayoutModule
+  ],
+  exports: [
+    HeaderComponent,
+    LayoutModule
+  ],
+  declarations: [
+    HeaderComponent
+  ]
+})
+export class HeaderModule {}