+++ projects/coreui-angular/src/lib/sidebar/sidebar-footer.component.ts
@@ -1,0 +1,17 @@
+import {Component, ElementRef, OnInit, Renderer2} from '@angular/core';
+
+@Component({
+  selector: 'cui-sidebar-footer',
+  template: `<ng-content></ng-content>`
+})
+export class SidebarFooterComponent implements OnInit {
+
+  constructor(
+    private renderer: Renderer2,
+    private hostElement: ElementRef
+  ) {
+    renderer.addClass(hostElement.nativeElement, 'app-sidebar-footer');
+  }
+
+  ngOnInit() { }
+}