+++ projects/coreui-angular/src/lib/sidebar/sidebar-form.component.ts
@@ -1,0 +1,16 @@
+import {Component, ElementRef, OnInit, Renderer2} from '@angular/core';
+
+@Component({
+  selector: 'cui-sidebar-form',
+  template: `<ng-content></ng-content>`
+})
+export class SidebarFormComponent implements OnInit {
+
+  constructor(
+    private renderer: Renderer2,
+    private hostElement: ElementRef
+  ) {
+    renderer.addClass(hostElement.nativeElement, 'sidebar-form');
+  }
+  ngOnInit() { }
+}