+++ projects/coreui-angular/src/lib/sidebar/sidebar-nav/sidebar-nav-divider.component.ts
@@ -1,0 +1,13 @@
+import {Component, Input, OnInit} from '@angular/core';
+
+@Component({
+  selector: 'cui-sidebar-nav-divider',
+  template: ``
+})
+export class SidebarNavDividerComponent implements OnInit {
+  @Input() item: any;
+
+  constructor () {}
+
+  ngOnInit() {}
+}