+++ src/polyfills.ts
@@ -33,6 +33,8 @@ import 'core-js/es6/regexp';
 import 'core-js/es6/map';
 import 'core-js/es6/weak-map';
 import 'core-js/es6/set';
+import 'core-js/es7/array';
+import 'core-js/es7/object';
 
 /** IE10 and IE11 requires the following for NgClass support on SVG elements */
 // import 'classlist.js';  // Run `npm install --save classlist.js`.
@@ -40,11 +42,12 @@ import 'core-js/es6/set';
 /** IE10 and IE11 requires the following for the Reflect API. */
 import 'core-js/es6/reflect';
 
+// for IE10
+// import 'mutationobserver-shim';
+
 
 /** Evergreen browsers require these. **/
 // Used for reflect-metadata in JIT. If you use AOT (and only Angular decorators), you can remove.
-
-
 
 /**
  * Web Animations `@angular/platform-browser/animations`
@@ -58,9 +61,11 @@ import 'core-js/es6/reflect';
  * user can disable parts of macroTask/DomEvents patch by setting following flags
  */
 
+ (window as any).__Zone_disable_timers = true;
  (window as any).__Zone_disable_requestAnimationFrame = true; // disable patch requestAnimationFrame
  (window as any).__Zone_disable_on_property = true; // disable patch onProperty such as onclick
  (window as any).__zone_symbol__BLACK_LISTED_EVENTS = ['scroll', 'mousemove']; // disable patch specified eventNames
+ (window as any).__Zone_enable_cross_context_check = true;
 
  /*
  * in IE/Edge developer tools, the addEventListener will also be wrapped by zone.js